Update patch set 1
Patch Set 1:

> I'm reluctant to accept any patches that aren't upstream already.
 > If it's upstream, which libpng release is in ?

This modifies code which is not part of standard libpng. (We already have modifications to libpng. In this case, our modifications have a bug.) This appears to fix b/18410359 (at least, it touches the code that causes the problem - I'd like to test to verify that though).

Patch-set: 1