Update patch set 1
Patch Set 1:

I'm not familiar with libpng, so it's unclear to me what the intent of the original patch was. I'm adding the original author and reviewers of the patch where the bug was introduced for review here.

It seems odd to not fail when inflate() fails. I think the additional check was intended to be after checking that inflate succeeded. So, I moved the check further down and added an explicit error message.

I was able to reproduce this issue with my Nexus 6 on build MRA58N with the png file attached from the bug on Google code (see commit msg). 

Here are the steps to reproduce:
 - Download the .png on the Nexus 6
 - Install a 3rd party .png viewer. I used Solid Explorer from the Play Store
 - Open the .png
 - Depending on the app, it might ANR or the image will not finish loading
 - pl.solidexplorer2 was pegging one CPU at 100% (indicating a hang)

This patch fixed the issue for me on a couple of Cyanogen devices.

I found a similar fix to this issue here as well:
https://android-review.googlesource.com/#/c/194680/

Thanks!

Patch-set: 1
Label: Verified=0